[CilkRTSCilkFor] Move code for lowering Tapir loops to calls to __cilkrts_cilk_for into a separate file, and make it accessible from multiple Tapir targets, specifically, CilkABI and CilkRABI.
