+++ llvm/include/llvm/Transforms/Tapir/CilkABI.h
@@ -15,10 +15,10 @@
 
 #include "llvm/IR/IRBuilder.h"
 #include "llvm/Transforms/Tapir/LoweringUtils.h"
-#include "llvm/Transforms/Tapir/TapirLoopInfo.h"
 
 namespace llvm {
 class Value;
+class TapirLoopInfo;
 
 class CilkABI : public TapirTarget {
   ValueToValueMapTy DetachCtxToStackFrame;
@@ -117,39 +117,6 @@ public:
 
   LoopOutlineProcessor *getLoopOutlineProcessor(const TapirLoopInfo *TL) const
     override final;
-};
-
-/// The RuntimeCilkFor loop-outline processor transforms an outlined Tapir loop
-/// to be processed using a call to a runtime method __cilkrts_cilk_for_32 or
-/// __cilkrts_cilk_for_64.
-class RuntimeCilkFor : public LoopOutlineProcessor {
-  FunctionCallee CilkRTSCilkFor32 = nullptr;
-  FunctionCallee CilkRTSCilkFor64 = nullptr;
-  Type *GrainsizeType = nullptr;
-
-  FunctionCallee Get__cilkrts_cilk_for_32();
-  FunctionCallee Get__cilkrts_cilk_for_64();
-public:
-  RuntimeCilkFor(Module &M) : LoopOutlineProcessor(M) {
-    GrainsizeType = Type::getInt32Ty(M.getContext());
-  }
-
-  ArgStructMode getArgStructMode() const override final {
-    // return ArgStructMode::Dynamic;
-    return ArgStructMode::Static;
-  }
-  void setupLoopOutlineArgs(
-      Function &F, ValueSet &HelperArgs, SmallVectorImpl<Value *> &HelperInputs,
-      ValueSet &InputSet, const SmallVectorImpl<Value *> &LCArgs,
-      const SmallVectorImpl<Value *> &LCInputs,
-      const ValueSet &TLInputsFixed)
-    override final;
-  unsigned getIVArgIndex(const Function &F, const ValueSet &Args) const
-    override final;
-  void postProcessOutline(TapirLoopInfo &TL, TaskOutlineInfo &Out,
-                          ValueToValueMapTy &VMap) override final;
-  void processOutlinedLoopCall(TapirLoopInfo &TL, TaskOutlineInfo &TOI,
-                               DominatorTree &DT) override final;
 };
 }  // end of llvm namespace
